sync: rcu: double linked list updates/reads with RCU
This example shows how to control concurrent updates with reads to a
double linked list. The synchronization mechanism used was a lockless
one: Read-Copy-Update (RCU).

Insertion was enabled to the userspace through sysfs, that allow user
processes fork lots of threads to concurrently update the data structure
within kernel space. The control for theses updates were made through
spin locks.

Deletion was made automatically through a kernel timer: every fixed
amount of time the timer expiration is hit and the first element of the
list is deleted with no conditions. (Also is controlled together with
insertion's lock)

Read are made through sysfs as well, where multiple user threads
(processes) can access the linked list to read its elements. The reads
occur concurrently to any update (insertion and/or deletion) and the
control is made through RCU.